Update 1.0
Chicken

+Name: string <<get/private set>>

+Hunger: int <<get/private set>>

+Happiness: int <<get/private set>>

+Eggs: int <<get/private set>>

+Chicken(string, int, int, int): void

+AdjustHunger(int): void

+AdjustHappiness(int): void

+MakeSound(): void

+Feed(string): void

+GetStatus(): void

+Sleep(): void

+Play(): void (not yet done)

+LayEgg(): void(not required yet)

-----------------------------------------------

Cow

+Name string <<get/private set>>

+Hunger: int <<get/private set>>

+Happiness: int <<get/ private set>>

+Milk: float <<get/private set>>

+Cow(string, int, int, float): void

+AdjustHunger(int): void

+AdjustHappiness(int): void

+MakeSound(): void

+Feed(string): void

+GetStatus(): void

+Moo(): void

+Play(): void (not required yet)

+ProduceMilk(): void(not yet done}

-----------------------------------------------

1) Write the Class Chicken and Cow that have the Attributes and Behavior as shown in the image below.

a. Hunger and Happiness values ​​are set to zero to 50.

b. Attributes: All variables must be treated as properties (Encapsulation).

c. The increase or decrease of Hunger and Happiness values ​​must be done only through the AdjustHunger and AdjustHappiness methods for data security (Encapsulation).

d. When food is given, it will result in animals feeling less hungry and being happier.

e. If a chicken gets sleep, it will slightly increase its Hunger and increase its Happiness.

f. If the cow mooes loudly (Moo), it will increase Happiness by 10 I.

g. every action in the game that causes Hunger and Happiness values ​​to change must display the changed values.

2) Create a GameManager class (or the Main class that we learned before, from now on we will use the name GameManager instead of Main).

3) In the GameManager class to run the Farm Sim game, do the following:

a. Create 1 Chicken and 1 Cow object from the Chicken and Cow classes.

b. Write a command to test the program to ensure that all methods work correctly.